Fix changingScraper errors (again)
The idea of this is to move the step where the new video is loaded (and the old video is accessed) to the main thread where ExoPlayer is to resolve some errors.

My process for testing was to break Option2 and fix Option1. Both Option1 and Option2 now work for vidstreaming.io properly. The changing servers works properly.

useFallback() would also have thrown an error before. This is also fixed in this commit.